# .spec/api-spec.yaml
openapi: "3.1.0"
info:
  title: Clara API
  version: "1.0.0"
  description: |
    Clara — AI Chat Receptionist for Local SMBs.

    ## Auth Model

    **Operator endpoints** require a static bearer token:
    `Authorization: Bearer <CLARA_OPERATOR_API_KEY>`

    **Visitor endpoints** require a valid session UUID in the request body or query params.
    The UUID is the capability token. No user authentication exists.

    ## Rate Limiting

    All endpoints enforce in-memory sliding window rate limits.
    Rate limit violations return HTTP 429 with `Retry-After: 60`.

    | Endpoint | Limit | Window |
    |----------|-------|--------|
    | POST /api/chat | 10 req/min/IP | 60s |
    | GET /api/chat | 30 req/min/IP | 60s |
    | POST /api/demo | 10 req/min/IP | 60s |
    | GET /api/demo | 30 req/min/IP | 60s |
    | POST /api/leads | 5 req/min/IP | 60s |
    | GET /api/leads | 20 req/min/IP | 60s |
    | DELETE /api/leads/{id} | 20 req/min/IP | 60s |
    | POST /api/admin/cleanup | 2 req/min/IP | 60s |
servers:
  - url: http://localhost:3002
    description: Local development
  - url: https://clara-staging.railway.app
    description: Staging (Railway)
  - url: https://clara.railway.app
    description: Production (Railway)
...
paths:
...
  /api/leads:
    post:
      summary: Capture a visitor lead
      description: |
        Records a visitor's contact information as a lead capture event.

        The `hubspot_company_id` is NEVER accepted from the client — it is read from
        the session row in the database. This prevents cross-tenant data injection.

        One session can have multiple lead rows (visitor submits more than once, or
        multiple people use the same link). No UNIQUE constraint on session_id.

        Rate limits:
        - 5 requests/minute/IP
        - 10 lead captures/session/lifetime (checked against DB count)
      operationId: captureLead
      tags: [leads]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureLeadRequest"
            example:
              sessionId: "550e8400-e29b-41d4-a716-446655440000"
              name: "James Chen"
              contact: "james@example.com"
              message: "Interested in beard trim pricing"
      responses:
        "201":
          description: Lead captured successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureLeadResponse"
              example:
                leadId: "770e8400-e29b-41d4-a716-446655440002"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Session not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "Session not found"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: List lead captures for a company (operator-only)
      description: |
        Returns all lead capture events for a given HubSpot company ID,
        ordered by most recent first.

        This endpoint is operator-only and requires bearer token auth.
        Lead data contains PII (name and contact info) — never return it without auth.
      operationId: getLeads
      tags: [leads]
      security:
        - operatorBearer: []
      parameters:
        - name: company
          in: query
          required: true
          description: The HubSpot company ID to query leads for
          schema:
            type: string
            maxLength: 64
          example: "123456789"
      responses:
        "200":
          description: Leads returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLeadsResponse"
              example:
                leads:
                  - id: "770e8400-e29b-41d4-a716-446655440002"
                    sessionId: "550e8400-e29b-41d4-a716-446655440000"
                    name: "James Chen"
                    contact: "james@example.com"
                    message: "Interested in beard trim pricing"
                    createdAt: "2026-03-24T10:20:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── DELETE /api/leads/{id} ───────────────────────────────────────────────
...
components:
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      description: |
        Static API key set in `CLARA_OPERATOR_API_KEY` environment variable.
        Generate with: `openssl rand -base64 32`
        Compared using `crypto.timingSafeEqual` — not string equality.
  schemas:
...
    CaptureLeadRequest:
      type: object
      required: [sessionId, name, contact]
      additionalProperties: false
      properties:
        sessionId:
          type: string
          format: uuid
          description: The session in which the lead capture was triggered
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Visitor's self-reported name. Free-text, not normalized.
          minLength: 1
          maxLength: 200
          example: "James Chen"
        contact:
          type: string
          description: |
            Visitor's email address or phone number. Single field — visitor provides
            whichever they prefer. Not validated as email or phone in v1.
          minLength: 1
          maxLength: 200
          example: "james@example.com"
        message:
          type: string
          description: Optional note from the visitor alongside their contact info
          maxLength: 1000
          example: "Interested in beard trim pricing"

    # ─── Response schemas ───────────────────────────────────────────────────
...
    CaptureLeadResponse:
      type: object
      required: [leadId]
      properties:
        leadId:
          type: string
          format: uuid
          description: UUID of the new lead record
          example: "770e8400-e29b-41d4-a716-446655440002"
    GetLeadsResponse:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          description: Lead capture events for the company, ordered by most recent first
          items:
            $ref: "#/components/schemas/LeadSummary"
    LeadSummary:
      type: object
      required: [id, sessionId, name, contact, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        sessionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "James Chen"
        contact:
          type: string
          description: Email or phone number provided by the visitor
          example: "james@example.com"
        message:
          type: string
          nullable: true
          example: "Interested in beard trim pricing"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-24T10:20:00.000Z"
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message. Never includes stack traces or internal IDs.
          example: "Session not found"
  responses:
    BadRequest:
      description: Invalid request — missing or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            missingField:
              summary: Missing required field
              value:
                error: "Missing required field: hubspot_company_id"
            invalidFormat:
              summary: Invalid field format
              value:
                error: "Invalid sessionId format"
            messageTooLong:
              summary: Message exceeds max length
              value:
                error: "Message too long (max 2000 characters)"
    Unauthorized:
      description: Missing or invalid operator API key
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: "Unauthorized"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
          example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            ipLimit:
              summary: Per-IP rate limit
              value:
                error: "Rate limit exceeded. Try again in 60 seconds."
            sessionHourly:
              summary: Per-session hourly limit
              value:
                error: "This session has reached its hourly message limit. Please try again in an hour."
            sessionLifetime:
              summary: Per-session lifetime cap
              value:
                error: "This demo has reached its message limit. Please contact us directly."
            leadSessionCap:
              summary: Per-session lead submission cap
              value:
                error: "Maximum lead submissions reached for this session."
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: "Internal server error"